requirements.txt
acres==0.2.0
aiofiles==24.1.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.7.0
attrs==24.3.0
cachetools==5.5.0
certifi==2024.12.14
charset-normalizer==3.4.0
ci-info==0.3.0
click==8.1.7
configobj==5.0.9
configparser==7.1.0
datasets==3.2.0
dill==0.3.8
distro==1.9.0
docstring_parser==0.16
etelemetry==0.3.1
filelock==3.16.1
frozenlist==1.5.0
fsspec==2024.9.0
google-ai-generativelanguage==0.6.10
google-api-core==2.24.0
google-api-python-client==2.156.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.3
googleapis-common-protos==1.66.0
grpcio==1.68.1
grpcio-status==1.68.1
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.27.2
huggingface-hub==0.27.0
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.4.5
instructor==1.7.0
isodate==0.6.1
Jinja2==3.1.4
jiter==0.6.1
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
litellm==1.55.6
looseversion==1.3.0
lxml==5.3.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
multidict==6.1.0
multiprocess==0.70.16
networkx==3.4.2
nibabel==5.3.2
nipype==1.9.2
numpy==2.2.0
openai==1.58.1
packaging==24.2
pandas==2.2.3
pathlib==1.0.1
pillow==11.0.0
propcache==0.2.1
proto-plus==1.25.0
protobuf==5.29.2
prov==2.0.1
puremagic==1.28
pyarrow==18.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.4
pydantic_core==2.27.2
pydot==3.0.3
Pygments==2.18.0
PyMuPDF==1.25.1
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
pyxnat==1.6.2
PyYAML==6.0.2
rdflib==6.3.2
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
rich==13.9.4
rpds-py==0.22.3
rsa==4.9
scikit-learn==1.6.0
scipy==1.14.1
shellingham==1.5.4
simplejson==3.19.3
six==1.17.0
sniffio==1.3.1
tenacity==9.0.0
threadpoolctl==3.5.0
tiktoken==0.8.0
tokenizers==0.21.0
tqdm==4.67.1
traits==6.4.3
typer==0.15.1
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
xxhash==3.5.0
yarl==1.18.3
zipp==3.21.0
